+++ ontimize-jee-report-rest/src/main/java/com/ontimize/jee/report/rest/ReportStoreRestController.java
@@ -1,24 +1,30 @@
 package com.ontimize.jee.report.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ontimize.jee.common.db.NullValue;
-import com.ontimize.jee.common.db.SQLStatementBuilder;
-import com.ontimize.jee.common.db.SQLStatementBuilder.BasicExpression;
 import com.ontimize.jee.common.dto.EntityResult;
 import com.ontimize.jee.common.dto.EntityResultMapImpl;
+import com.ontimize.jee.report.common.dto.ReportStoreParamValueDto;
+import com.ontimize.jee.report.common.dto.ReportStoreParamsDto;
 import com.ontimize.jee.report.common.exception.ReportStoreException;
 import com.ontimize.jee.report.common.reportstore.BasicReportDefinition;
 import com.ontimize.jee.report.common.reportstore.ReportOutputType;
 import com.ontimize.jee.report.common.reportstore.ReportParameter;
 import com.ontimize.jee.report.common.services.IReportDefinition;
 import com.ontimize.jee.report.common.services.IReportStoreService;
-import com.ontimize.jee.server.rest.BasicExpressionProcessor;
+import com.ontimize.jee.report.common.util.TypeMappingsUtils;
+import com.ontimize.jee.server.rest.FilterParameter;
 import com.ontimize.jee.server.rest.ORestController;
 import com.ontimize.jee.server.rest.ParseUtilsExt;
+import com.ontimize.jee.server.rest.QueryParameter;
+import org.apache.commons.compress.utils.FileNameUtils;
+import org.apache.commons.io.FileUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.MediaType;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,13 +36,9 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.sql.Date;
-import java.sql.Timestamp;
-import java.sql.Types;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
@@ -45,7 +47,9 @@ import java.util.concurrent.ExecutionException;
 @RestController
 @RequestMapping("/reportstore")
 @ComponentScan(basePackageClasses = {IReportStoreService.class})
-public class ReportStoreRestController {
+public class ReportStoreRestController extends ORestController<IReportStoreService> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ReportStoreRestController.class);
 
     @Qualifier("ReportStoreService")
     @Autowired
@@ -69,61 +73,52 @@ public class ReportStoreRestController {
             }
 
             String id = UUID.randomUUID().toString();
-            String mainReportFilename = files[0].getOriginalFilename().split("\\.")[0] + ".jrxml";
+            if(files == null || !(files.length > 0) || files[0] == null ){
+                throw new IOException("No report file found");
+            }
+            MultipartFile multipartFile = files[0];
+            String fileName = StringUtils.isEmpty(multipartFile.getOriginalFilename()) ? "reportFile" : multipartFile.getOriginalFilename();
+            String mainReportFilename = FileNameUtils.getBaseName(fileName) + ".jrxml";
 
             IReportDefinition rdef = new BasicReportDefinition(id, extraData.get("name").toString(), extraData.get("description").toString(),
                     extraData.get("type").toString(), mainReportFilename);
-            InputStream reportSource = new ByteArrayInputStream(files[0].getBytes());
+            InputStream reportSource = new ByteArrayInputStream(multipartFile.getBytes());
 
             return this.reportStoreService.addReport(rdef, reportSource);
-        } catch (ReportStoreException e) {
-            e.printStackTrace();
+        } catch (ReportStoreException | IOException e) {
+            LOGGER.error("Error adding report", e);
             res.setCode(EntityResult.OPERATION_WRONG);
+            res.setMessage("Error adding report! " + e.getMessage());
             return res;
-        } catch (IOException e) {
-            e.printStackTrace();
-            res.setCode(EntityResult.OPERATION_WRONG);
-            return res;
-        }
+        } 
     }
 
     @SuppressWarnings("unchecked")
     @RequestMapping(value = "/fillReport/{id}", method = RequestMethod.POST)
     public EntityResult fillReport(@PathVariable("id") String id,
-                                   @RequestBody(required = true) Map<String, Object> bodyParams) {
+                                   @RequestBody(required = true) ReportStoreParamsDto bodyParams) {
         EntityResult res = new EntityResultMapImpl();
         ReportOutputType outputType;
         String otherType = "pdf";
-        String[] values;
-        Map<Object, Object> filter = ((Map<Object, Object>) bodyParams.get("filter"));
+
         Map<Object, Object> keysValues = new HashMap<Object, Object>();
         try {
             outputType = ReportOutputType.fromName("pdf");
-            Map<String, Object> params = new HashMap<String, Object>();
-            if (!((String) bodyParams.get("params")).isEmpty()) {
-                IReportDefinition rDef = this.parseReportEntityResult(this.reportStoreService.getReportDefinition(id));
-                values = ((String) bodyParams.get("params")).split("\\,");
-                for (int i = 0; i < rDef.getParameters().size(); i++) {
-                    params.put(rDef.getParameters().get(i).getName(), this.parseParameter(rDef, i, values[i]));
-                }
-            }
-            if (!filter.isEmpty()) {
-                filter = (Map<Object, Object>) filter.get("filter");
-                keysValues = this.createKeysValues(filter, new HashMap<>());
+            Map<String, Object> params = processReportParameters(id, bodyParams);
+            if (bodyParams.getFilters() != null) {
+                keysValues = processFilterParameter(bodyParams).getFilters().getFilter();
             }
 
             CompletableFuture<EntityResult> future = this.reportStoreService.fillReport(id, params, null, outputType, otherType, keysValues);
             res = future.get();
-        } catch (ReportStoreException e) {
-            e.printStackTrace();
+        } catch (ReportStoreException | InterruptedException | ExecutionException e) {
+            LOGGER.error("Error filling report", e);
+            if(e instanceof InterruptedException) {
+                Thread.currentThread().interrupt();
+            }
             res.setCode(EntityResult.OPERATION_WRONG);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-            res.setCode(EntityResult.OPERATION_WRONG);
-        } catch (ExecutionException e) {
-            e.printStackTrace();
-            res.setCode(EntityResult.OPERATION_WRONG);
-        }
+            res.setMessage("Error filling report! " + e.getMessage());
+        } 
 
         return res;
     }
@@ -134,8 +129,9 @@ public class ReportStoreRestController {
         try {
             return this.reportStoreService.removeReport(id);
         } catch (ReportStoreException e) {
-            e.printStackTrace();
+            LOGGER.error("Error removing report", e);
             res.setCode(EntityResult.OPERATION_WRONG);
+            res.setMessage("Error removing report! " + e.getMessage());
             return res;
         }
     }
@@ -146,8 +142,9 @@ public class ReportStoreRestController {
         try {
             return this.reportStoreService.listAllReports();
         } catch (ReportStoreException e) {
-            e.printStackTrace();
+            LOGGER.error("Error listing reports", e);
             res.setCode(EntityResult.OPERATION_WRONG);
+            res.setMessage("Error listing reports! " + e.getMessage());
             return res;
         }
     }
@@ -158,8 +155,9 @@ public class ReportStoreRestController {
         try {
             return this.reportStoreService.getReportDefinition(id);
         } catch (ReportStoreException e) {
-            e.printStackTrace();
+            LOGGER.error("Error retrieving report", e);
             res.setCode(EntityResult.OPERATION_WRONG);
+            res.setMessage("Error retrieving report! " + e.getMessage());
             return res;
         }
     }
@@ -186,8 +184,9 @@ public class ReportStoreRestController {
 
             return this.reportStoreService.updateReportDefinition(rdef);
         } catch (ReportStoreException e) {
-            e.printStackTrace();
+            LOGGER.error("Error updating report", e);
             res.setCode(EntityResult.OPERATION_WRONG);
+            res.setMessage("Error updating report! " + e.getMessage());
             return res;
         }
     }
@@ -222,83 +221,56 @@ public class ReportStoreRestController {
         return map;
     }
 
-    private Object parseParameter(IReportDefinition rDef, int index, String value) {
-        String type = rDef.getParameters().get(index).getValueClass();
-        switch (type) {
-            case "java.lang.Integer":
-                return Integer.parseInt(value);
-            case "java.lang.Double":
-                return Double.parseDouble(value);
-            case "java.lang.Float":
-                return Double.parseDouble(value);
-            case "java.lang.Real":
-                return Float.parseFloat(value);
-            case "java.sql.Date":
-                return Date.valueOf(value);
-            case "java.sql.Timestamp":
-                return Timestamp.valueOf(value);
-            case "java.lang.String":
-                break;
+    private Object parseParameter(ReportParameter reportParameter, ReportStoreParamValueDto paramValueDto) {
+        Integer sqlType = null;
+        if(paramValueDto.getSqlType() != null) {
+            sqlType = paramValueDto.getSqlType();
+        } else {
+            sqlType = TypeMappingsUtils.getSQLTypeFromClassName(reportParameter.getValueClass());
         }
-
-        return value;
+        return ParseUtilsExt.getValueForSQLType(paramValueDto.getValue(), sqlType.intValue());
     }
-
-    protected Map<Object, Object> createKeysValues(Map<?, ?> kvQueryParam, Map<?, ?> hSqlTypes) {
-        Map<Object, Object> kv = new HashMap<>();
-        if ((kvQueryParam == null) || kvQueryParam.isEmpty()) {
-            return kv;
-        }
-
-        if (kvQueryParam.containsKey(ORestController.BASIC_EXPRESSION)) {
-            Object basicExpressionValue = kvQueryParam.remove(ORestController.BASIC_EXPRESSION);
-            this.processBasicExpression(SQLStatementBuilder.ExtendedSQLConditionValuesProcessor.EXPRESSION_KEY, kv,
-                    basicExpressionValue, hSqlTypes);
-        }
-
-        if (kvQueryParam.containsKey(ORestController.FILTER_EXPRESSION)) {
-            Object basicExpressionValue = kvQueryParam.remove(ORestController.FILTER_EXPRESSION);
-            this.processBasicExpression(SQLStatementBuilder.ExtendedSQLConditionValuesProcessor.FILTER_KEY, kv,
-                    basicExpressionValue, hSqlTypes);
-        }
-
-        for (Entry<?, ?> next : kvQueryParam.entrySet()) {
-            Object key = next.getKey();
-            Object value = next.getValue();
-            if ((hSqlTypes != null) && hSqlTypes.containsKey(key)) {
-                int sqlType = (Integer) hSqlTypes.get(key);
-                value = ParseUtilsExt.getValueForSQLType(value, sqlType);
-                if (value == null) {
-                    if (ParseUtilsExt.BASE64 == sqlType) {
-                        sqlType = Types.BINARY;
-                    }
-                    value = new NullValue(sqlType);
+    
+    protected Map<String, Object> processReportParameters(final String id, final ReportStoreParamsDto paramsDto) throws ReportStoreException {
+        Map<String, Object> params = new HashMap<String, Object>();
+        if (paramsDto.getParameters() != null && !paramsDto.getParameters().isEmpty()) {
+            IReportDefinition rDef = this.parseReportEntityResult(this.reportStoreService.getReportDefinition(id));
+            List<ReportParameter> reportParameters = rDef.getParameters();
+            List<ReportStoreParamValueDto> params1 = paramsDto.getParameters();
+            for (int i = 0; i < reportParameters.size(); i++) {
+                ReportParameter currentRepParam = reportParameters.get(i);
+                ReportStoreParamValueDto paramValueDto = params1.stream().filter(item -> item.getName().equals(currentRepParam.getName())).findFirst().orElse(null);
+                if(paramValueDto != null) {
+                    params.put(currentRepParam.getName(), this.parseParameter(currentRepParam, paramValueDto));
                 }
-            } else if (value == null) {
-                value = new NullValue();
-            }
-            kv.put(key, value);
-        }
-
-        return kv;
-    }
-
-    @SuppressWarnings("unchecked")
-    protected void processBasicExpression(String key, Map<?, ?> keysValues, Object basicExpression,
-                                          Map<?, ?> hSqlTypes) {
-        if (basicExpression instanceof Map) {
-            try {
-                BasicExpression bE = BasicExpressionProcessor.getInstance()
-                        .processBasicEspression(basicExpression, hSqlTypes);
-                ((Map<Object, Object>) keysValues).put(key, bE);
-            } catch (Exception e) {
-                e.printStackTrace();
             }
         }
-
+        return params;
     }
 
-    protected void processBasicExpression(String key, Map<Object, Object> keysValues, Object basicExpression) {
-        this.processBasicExpression(key, keysValues, basicExpression, new HashMap<>());
+    protected ReportStoreParamsDto processFilterParameter(final ReportStoreParamsDto param) throws ReportStoreException {
+        if (param == null) {
+            throw new ReportStoreException("'ReportStoreParams' not found!");
+        }
+
+        FilterParameter filterParam = param.getFilters();
+        if (filterParam == null) {
+            filterParam = new QueryParameter();
+        }
+
+        Map<?, ?> kvQueryParameter = filterParam.getFilter();
+        List<?> avQueryParameter = filterParam.getColumns();
+        Map<?, ?> hSqlTypes = filterParam.getSqltypes();
+
+        Map<Object, Object> processedKeysValues = this.createKeysValues(kvQueryParameter, hSqlTypes);
+        List<Object> processedAttributesValues = this.createAttributesValues(avQueryParameter, hSqlTypes);
+
+        filterParam.setKv(processedKeysValues);
+        filterParam.setColumns(processedAttributesValues);
+
+        param.setFilters(filterParam);
+
+        return param;
     }
+
 }